refactor(listings): implement CQRS query pattern and domain value objects
- Added Domain Value Objects for listings: Title, Price, Model, YearOfIssue.
- Integrated Query Services into UseCases, decoupling read and write logic (CQRS).
- Optimized dependencies and imports in listings application layer.
- Removed AbstractListingReferenceUnitOfWork in favor of ReferenceQueryContext.

Closes #52 #53
