+++ src/core/listings/application/abstract_listing_query_context.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod
+
+
+class AbstractListingQueryContext(ABC):
+    @property
+    @abstractmethod
+    def listing(self):
+        raise NotImplementedError
+
+    @property
+    @abstractmethod
+    def filter(self):
+        raise NotImplementedError
