+++ src/core/listings/application/abstract_listing_unit_of_work.py
@@ -1,0 +1,8 @@
+from abc import abstractmethod, ABC
+
+
+class AbstractListingUnitOfWork(ABC):
+    @property
+    @abstractmethod
+    def listing(self):
+        raise NotImplementedError
