+++ src/core/listings/application/usecases/create_draft_listing.py
@@ -2,11 +2,11 @@ from uuid import UUID
 
 from src.api.listings.dto import CreateDraftListingDTO
 from src.core.listings.infrastructure.factories import ListingFactory
-from src.core.shared.application.abstract_uow import AbstractListingReferenceUnitOfWork
+from src.core.listings.application.abstract_listing_unit_of_work import AbstractListingUnitOfWork
 
 
 class CreateDraftListingUseCase:
-    def __init__(self, unit_of_work: AbstractListingReferenceUnitOfWork):
+    def __init__(self, unit_of_work: AbstractListingUnitOfWork):
         self.unit_of_work = unit_of_work
 
     async def execute(self, draft_listing_dto: CreateDraftListingDTO, author_id: UUID):
