+++ src/core/listings/application/usecases/get_detail_public_listing.py
@@ -1,14 +1,14 @@
 from uuid import UUID
 
-from src.core.shared.application.abstract_uow import AbstractListingReferenceUnitOfWork
+from src.core.listings.application.abstract_listing_query_context import AbstractListingQueryContext
 
 
 class GetDetailPublicListingUseCase:
-    def __init__(self, unit_of_work: AbstractListingReferenceUnitOfWork):
-        self.unit_of_work = unit_of_work
+    def __init__(self, query_service: AbstractListingQueryContext):
+        self.query_service = query_service
 
     async def execute(self, listing_id: UUID):
-        async with self.unit_of_work as uow:
-            listing = await uow.listing_query.get_detail_listing_by_id(listing_id)
+        async with self.query_service as query:
+            listing = await query.listing.get_detail_listing_by_id(listing_id)
             return listing
