+++ src/core/listings/infrastructure/query_services/form_query.py
@@ -1,0 +1,6 @@
+from sqlalchemy.ext.asyncio import AsyncSession
+
+
+class ListingFormQueryService:
+    def __init__(self, session: AsyncSession):
+        self._session = session
