+++ src/core/references/application/abstract_reference_query_context.py
@@ -1,0 +1,18 @@
+from abc import ABC, abstractmethod
+
+
+class AbstractReferenceQueryContext(ABC):
+    @property
+    @abstractmethod
+    def reference(self):
+        raise NotImplementedError
+
+    @property
+    @abstractmethod
+    def category(self):
+        raise NotImplementedError
+
+    @property
+    @abstractmethod
+    def specification(self):
+        raise NotImplementedError
