--- src/core/references/application/abstract_reference_uow.py
@@ -1,11 +1,0 @@
-from abc import abstractmethod
-
-from src.core.shared.application.abstract_uow import AbstractUnitOfWork
-
-
-class AbstractReferenceUnitOfWork(AbstractUnitOfWork):
-    @property
-    @abstractmethod
-    def reference(self):
-        raise NotImplementedError
-
